+++ src/presenter/trip-presenter.js
@@ -2,7 +2,8 @@ import AddFirstPoint from '../view/site-add-first-point';
 import EventsListTemplate from '../view/site-list-view';
 import PointPresenter from './point-presenter';
 import { render, RenderPosition, remove } from '../utils/render';
-import { SortType, sortDate, sortTime, sortPrice } from '../utils/informations';
+import { sortPrice, SortType } from '../utils/common';
+import { sortDate, sortTime } from '../utils/functionsWithDayjs';
 import { UpdateType } from '../const';
 import { FilterType, UserAction } from '../const';
 import EventNewPresenter from './point-new-presenter';
